+++ src/backend/core/tests/test_api_users.py
@@ -48,7 +48,7 @@ def test_api_users_list_authenticated():
 
 def test_api_users_list_query_inactive():
     """Inactive users should not be listed."""
-    user = factories.UserFactory()
+    user = factories.UserFactory(email="alice@example.com")
     client = APIClient()
     client.force_login(user)
 
@@ -67,7 +67,7 @@ def test_api_users_list_query_short_queries():
     Queries shorter than 5 characters should return an empty result set.
     """
 
-    user = factories.UserFactory()
+    user = factories.UserFactory(email="alice@example.com")
     client = APIClient()
     client.force_login(user)